+++ src/main/java/baseball/controller/GameManager.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.controller;
 
 import baseball.domain.BallNumber;
