+++ src/main/java/baseball/controller/NumberComparator.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.controller;
 
 public class NumberComparator {
     static boolean isEqual(int computerNumber, int playerNumber) {
