+++ src/test/java/baseball/NumberComparatorTest.java
@@ -3,6 +3,7 @@ package baseball;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import baseball.controller.NumberComparator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
